requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Database (for production)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Supabase
supabase==2.0.2
postgrest==0.12.0

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# Fix for pkg_resources issue
setuptools>=65.0.0

# Image processing and PDF generation (Python 3.11 compatible versions)
Pillow>=9.0.0,<11.0.0
qrcode[pil]==7.4.2
reportlab==4.0.7
pdfkit==1.0.0

# Excel file processing
openpyxl==3.1.2

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
# WeasyPrint==60.1               # Alternative PDF renderer
weasyprint==60.1
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
# pandas==2.1.3                  # Heavy data processing (bulk ops)
# python-dateutil==2.8.2         # Date utilities